Fix some relative path links in the docs (#215)
While browsing the docs (and then when running `just doc`) I found some
broken links.

There's still a warning when running `just doc`:
```
WARNING -  A reference to 'quick_start/qourier/index.md' is included in the 'nav' configuration, which is not found in the documentation files.
```
But that's just due to the entry in the nav, so I ignored it for now.