Add nightly CI for documentation builds (#231)
Closes #228 and possibly also #219 

Needs a release of bloqade-circuit v0.2.0 first as it also updates the
examples to work with the new version.

Also, I'm checking the versions / commit hashes for the nightly builds
and cancel the jobs if nothing has changed. To do this, I'm using github
variables. Not sure if this is the best solution. If it is, then we'd
need to create the variables (the body can be empty to start with).
There are probably permission issues then as well since I don't think
the `GH_TOKEN` has permissions to set github variables and cancel jobs.

@Roger-luo @weinbe58 do you know a better solution for this?