remove bloqade-pyqrack mention (#261)
As far as I understand bloqade-pyqrack is no longer being developed and
we shouldn't recommend folks install it. Might be worth updating the
readme on bloqade-pyqrack to make this *very* explicit.

I've also added a little note that there's no need to install the
subpackages if you just do `pip install bloqade`. I know saying Bloqade
is a namespace package should already clarify that but I don't think it
hurts to make it more clear (: