pyqrack installation revisions and install page suggestions (#263)
Made the install page a bit more cohesive as well as removing a mention
of `bloqade-pyqrack` considering it's now deprecated and users just need
to install `pyqrack` by itself.

I also added an explicit warning about installing `pyqrack-cpu` for
Intel Macs because @lamq317 ran into this issue on his machine and it
didn't seem like we had it documented anywhere.

I think it might even be worth bringing up this pyqrack install info on
the bloqade install page instead of having it tucked away in its current
state in an "under construction" page.